LayoutTests/fast/dom/MicroData/domsettabletokenlist-attributes-out-of-range-index.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div itemtype="" id="mydiv"></div>
<script>
description("This test verifies itemType[index], itemRef[index], itemProp[index] behavior for out-of-range index.");

debug("Test itemType[index]");
var element = document.getElementById("mydiv");
shouldBeNull("element.itemType.item(0)");
shouldBeUndefined("element.itemType[0]");
element.itemType.add("foo");
shouldBeNull("element.itemType.item(1)");
shouldBeUndefined("element.itemType[1]");
element.itemType.remove("foo");
shouldBeNull("element.itemType.item(0)");
shouldBeUndefined("element.itemType[0]");
//Test with invalid index
shouldBeNull("element.itemType.item(-1)");
shouldBeUndefined("element.itemType[-1]");

debug("\nTest itemRef[index]");
var element = document.getElementById("mydiv");
shouldBeNull("element.itemRef.item(0)");
shouldBeUndefined("element.itemRef[0]");
element.itemRef.add("foo");
shouldBeNull("element.itemRef.item(1)");
shouldBeUndefined("element.itemRef[1]");
element.itemRef.remove("foo");
shouldBeNull("element.itemRef.item(0)");
shouldBeUndefined("element.itemRef[0]");
//Test with invalid index
shouldBeNull("element.itemRef.item(-1)");
shouldBeUndefined("element.itemRef[-1]");

debug("\nTest itemProp[index]");
var element = document.getElementById("mydiv");
shouldBeNull("element.itemProp.item(0)");
shouldBeUndefined("element.itemProp[0]");
element.itemProp.add("foo");
shouldBeNull("element.itemProp.item(1)");
shouldBeUndefined("element.itemProp[1]");
element.itemProp.remove("foo");
shouldBeNull("element.itemProp.item(0)");
shouldBeUndefined("element.itemProp[0]");
//Test with invalid index
shouldBeNull("element.itemProp.item(-1)");
shouldBeUndefined("element.itemProp[-1]");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
